feat(model): support initial value field for boxes
This means that puzzles with initial values in boxes can be exported
without their save state with the initial values carried through. I
currently have no plans to actually render the initial value on the
board -- the readers prefill the response with the initial value as
well (which should then be overwritten with saved state if it exists).