add(io): parse arrow elements in cell JPZ
I don't fully understand the spec, so just use the "to" attribute to get
the arrow direction. I'm not sure what the from direction should do. I
guess an arrow from "topleft" to "right" should be an arrow along the
top, but all arrows go along the middle in Forkyz at the moment.