bug(downloaders): make notretempsgeant2 fortnightly
This is actually not a fix as Notre Temps is irregular. The code was
written thinking it was every two weeks, with some resulting
improvements to the abstract downloaders (allowing downloaders with gaps
more than one week). So this commit is to keep those changes.

Notre Temps should be a PageScraper. Will delete in the next commit
pending a rewrite.