bug(ime): honour batch edits and don't update until all done
Some keyboards report the replacement of the selection as a delete and
then an insert, rather than one replace. Done as a batch edit.

Forkyz should hold off on updating after the delete because that would
move the cursor backwards as if only delete had been pressed.

By waiting for the full edit to finish, it sees that it was a replace
and not just a delete.

Part of #70