bug(ui): close browse activity immediately after external import
Either ACTION_VIEW or ACTION_SEND. This is because they both mess with
ACTION_MAIN. After view or send, attempts to return to a puzzle from the
launcher create a new browse activity and exit the play activity. This
happens only once with view, but every time with send (until the app is
force closed and restarted).

Instead, after one of these actions, display a dialog to the user that
the import succeeded, then exit. They then have to start Forkyz from the
launcher and behaviour is as you would expect.

Solutions such as singleTask don't work as they always return to browse
when opened from the launcher. This is not a fun play experience.